fix the comment
