+++ backend/app/repositories/models/custom_bot.py
@@ -3,24 +3,9 @@ from app.repositories.models.common import Float
 from app.repositories.models.custom_bot_guardrails import BedrockGuardrailsModel
 from app.repositories.models.custom_bot_kb import BedrockKnowledgeBaseModel
 from app.routes.schemas.bot import type_sync_status
-from pydantic import BaseModel, ConfigDict, create_model
+from pydantic import BaseModel, ConfigDict
 from app.routes.schemas.conversation import type_model_name
-
-
-class DynamicBaseModel(BaseModel):
-    model_config = ConfigDict(extra="allow")
-
-
-def create_model_activate_model(model_names: List[str]) -> Type[DynamicBaseModel]:
-    fields: Dict[str, Any] = {
-        name.replace("-", "_").replace(".", "_"): (bool, False) for name in model_names
-    }
-    return create_model("ModelActivateModel", __base__=DynamicBaseModel, **fields)
-
-
-ModelActivateModel: Type[BaseModel] = create_model_activate_model(
-    list(get_args(type_model_name))
-)
+from app.repositories.models.common import ModelActivateModel
 
 
 class KnowledgeModel(BaseModel):
